// app/global/types.tsx
export interface cardProps {
  title: string;
  onPress?: any;
  bgColor?: string;
  buttonRadius?: number;
  buttonPadding?: number;
  buttonWidth?: any;
  buttonHeight?: number;
  buttonBorderWidth?: number;
  btnTextColor?: string;
}

export interface Props {}

export interface boxProps {
  id: number;
  value?: any;
}